requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.100.0
groq==0.4.2
